atf/requirements.txt
playwright==1.44.0
pytest==8.2.0
pytest-playwright==0.5.0
gitpython==3.1.43
pyyaml==6.0.1
click==8.1.7
rich==13.7.1
